+++ test/cldr_test.exs
@@ -49,7 +49,7 @@ defmodule Cldr.Test do
                cldr_locale_name: "en-001",
                language_subtags: [],
                extensions: %{},
-               gettext_locale_name: nil,
+               gettext_locale_name: "en",
                language: "en",
                locale: %{},
                private_use: [],