+++ test/plug/plug_set_locale_test.exs
@@ -21,7 +21,7 @@ defmodule Cldr.Plug.SetLocale.Test do
                cldr_locale_name: "en-001",
                language_subtags: [],
                extensions: %{},
-               gettext_locale_name: nil,
+               gettext_locale_name: "en",
                language: "en",
                locale: %{},
                private_use: [],